+++ docs/changes.rst
@@ -9,7 +9,8 @@ Released on ? 2017.
 * Improved admin interface for groups.
 * Added support for Yandex Translate API.
 * Improved speed of sitewide search.
-* Add project wide search.
+* Added project wide search.
+* Added project wide search and replace.
 
 weblate 2.11
 ------------